C21__C30_Projects/C26_AccessControlSystem_26_project/requirements.txt
# Django
Django==5.0.1
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# CORS
django-cors-headers==4.3.1

# Admin UI
django-jazzmin==2.6.0

# Static files
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.27.0

# Server
gunicorn==21.2.0

# Development
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0